datablock/requirements.txt
flask
flask-login
boto3
python-dotenv
cryptography
web3
py-solc-x
requests
